pkgs/tasks/pyproject.toml
[project]
name = "tasks"
version = "0.1.0"
description = "YAML-backed task store for Ralph."
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0",
]

[tool.setuptools]
packages = ["tasks"]

[tool.setuptools.package-dir]
tasks = "."
